refactor: domain 패키지 depth 통일
