+++ src/main/kotlin/com/yapp/yapp/record/domain/record/RunningRecord.kt
@@ -1,10 +1,10 @@
 package com.yapp.yapp.record.domain.record
 
 import com.yapp.yapp.common.TimeProvider
-import com.yapp.yapp.running.Pace
-import com.yapp.yapp.running.RecordStatus
-import com.yapp.yapp.running.converter.DurationConverter
-import com.yapp.yapp.running.converter.PaceConverter
+import com.yapp.yapp.running.domain.Pace
+import com.yapp.yapp.running.domain.RecordStatus
+import com.yapp.yapp.running.domain.converter.DurationConverter
+import com.yapp.yapp.running.domain.converter.PaceConverter
 import jakarta.persistence.Column
 import jakarta.persistence.Convert
 import jakarta.persistence.Entity
