+++ src/main/kotlin/com/yapp/yapp/running/domain/RecordStatus.kt
@@ -1,4 +1,4 @@
-package com.yapp.yapp.running
+package com.yapp.yapp.running.domain
 
 enum class RecordStatus {
     READY,
