+++ src/test/kotlin/com/yapp/yapp/running/RunningMetricsCalculatorTest.kt
@@ -3,6 +3,7 @@ package com.yapp.yapp.running
 import com.yapp.yapp.common.TimeProvider
 import com.yapp.yapp.record.domain.point.RunningPoint
 import com.yapp.yapp.record.domain.record.RunningRecord
+import com.yapp.yapp.running.domain.RunningMetricsCalculator
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
